+++ rational/src/helper_macros.rs
@@ -48,6 +48,92 @@ macro_rules! impl_binop_with_macro {
     };
 }
 
+/// Implement `impl Op<B> for A` by calling `$impl`. The arguments for this
+/// macro will be `(self.numerator, self.denominator, other,
+/// &self.numerator, &self.denominator, &other, $method)`.
+macro_rules! impl_binop_with_int {
+    (impl $trait:ident<$int:ty>, $method:ident, $impl:ident) => {
+        crate::helper_macros::impl_binop_with_int!(impl $trait<$int>, $method, crate::rbig::RBig, $impl);
+    };
+    (impl $trait:ident<$int:ty>, $method:ident, $t:ty, $impl:ident) => {
+        impl $trait<$int> for $t {
+            type Output = $t;
+            fn $method(self, rhs: $int) -> $t {
+                let (a, b) = self.into_parts();
+                let (ra, rb, ri) = (&a, &b, &rhs);
+                $impl!(a, b, rhs, ra, rb, ri, $method)
+            }
+        }
+
+        impl<'r> $trait<&'r $int> for $t {
+            type Output = $t;
+            fn $method(self, rhs: &$int) -> $t {
+                let (a, b) = self.into_parts();
+                let (ra, rb, ri) = (&a, &b, rhs);
+                $impl!(a, b, rhs, ra, rb, ri, $method)
+            }
+        }
+
+        impl<'l> $trait<$int> for &'l $t {
+            type Output = $t;
+            fn $method(self, rhs: $int) -> $t {
+                let (ra, rb, ri) = (self.numerator(), self.denominator(), &rhs);
+                let (a, b) = (ra.clone(), rb.clone());
+                $impl!(a, b, rhs, ra, rb, ri, $method)
+            }
+        }
+
+        impl<'l, 'r> $trait<&'r $int> for &'l $t {
+            type Output = $t;
+            fn $method(self, rhs: &$int) -> $t {
+                let (ra, rb, ri) = (self.numerator(), self.denominator(), rhs);
+                let (a, b) = (ra.clone(), rb.clone());
+                $impl!(a, b, rhs, ra, rb, ri, $method)
+            }
+        }
+    };
+    (impl $trait:ident for $int:ty, $method:ident, $impl:ident) => {
+        crate::helper_macros::impl_binop_with_int!(impl $trait for $int, $method, crate::rbig::RBig, $impl);
+    };
+    (impl $trait:ident for $int:ty, $method:ident, $t:ty, $impl:ident) => {
+        impl $trait<$t> for $int {
+            type Output = $t;
+            fn $method(self, rhs: $t) -> $t {
+                let (a, b) = rhs.into_parts();
+                let (ra, rb, ri) = (&a, &b, &self);
+                $impl!(a, b, self, ra, rb, ri, $method)
+            }
+        }
+
+        impl<'r> $trait<&'r $t> for $int {
+            type Output = $t;
+            fn $method(self, rhs: &$t) -> $t {
+                let (ra, rb, ri) = (rhs.numerator(), rhs.denominator(), &self);
+                let (a, b) = (ra.clone(), rb.clone());
+                $impl!(a, b, self, ra, rb, ri, $method)
+            }
+        }
+
+        impl<'l> $trait<$t> for &'l $int {
+            type Output = $t;
+            fn $method(self, rhs: $t) -> $t {
+                let (a, b) = rhs.into_parts();
+                let (ra, rb, ri) = (&a, &b, self);
+                $impl!(a, b, self, ra, rb, ri, $method)
+            }
+        }
+
+        impl<'l, 'r> $trait<&'r $t> for &'l $int {
+            type Output = $t;
+            fn $method(self, rhs: &$t) -> $t {
+                let (ra, rb, ri) = (rhs.numerator(), rhs.denominator(), self);
+                let (a, b) = (ra.clone(), rb.clone());
+                $impl!(a, b, self, ra, rb, ri, $method)
+            }
+        }
+    }
+}
+
 /// Implement `impl OpAssign<B> for A` by forwarding to `*A = mem::take(A).op(B)`, including &B.
 macro_rules! impl_binop_assign_by_taking {
     (impl $trait:ident<$t2:ty> for $t1:ty, $methodassign:ident, $method:ident) => {
@@ -68,3 +154,4 @@ macro_rules! impl_binop_assign_by_taking {
 
 pub(crate) use impl_binop_assign_by_taking;
 pub(crate) use impl_binop_with_macro;
+pub(crate) use impl_binop_with_int;